Move work with socket outside from main.js file (Perhaps better create some service etc) Handle submit form.
