+++ index.js
@@ -2,12 +2,10 @@ require('dotenv-safe').config();
 
 const bodyParser        = require('body-parser');
 const routes            = require('./routes');
+const handleSocket      = require('./controllers/socketController');
 const {app, Server, io} = require('./config');
 const express           = require('express')
-const twilioPhoneNumber = process.env.TWILIO_PHONE_NUMBER;
-const twilio            = require('twilio');
 const port              = process.env.APP_PORT;
-const WhatsAppMessage   = require('./models/whatsapp-message');
 
 /** Server Config */
 app.use(bodyParser.urlencoded({ extended: false }));
@@ -21,35 +19,7 @@ app.use('/', routes);
 
 /** Socket io connection */
 io.on('connection', (socket) => {
-  socket.on('wa_reply', function(msg) {
-    const accountId           = process.env.TWILIO_ACCOUNT_ID;
-    const authToken           = process.env.TWILIO_AUTH_TOKEN;
-    const client              = twilio(accountId, authToken);
-
-    client.messages.create({
-      from: twilioPhoneNumber,
-      body: msg.text,
-      to: 'whatsapp:' + msg.userId
-    }).then(response => {
-
-      console.log(response);
-      var newWhatsAppMessage = new WhatsAppMessage({
-        'From': response.to,
-        'Body': response.body
-      });
-
-      newWhatsAppMessage.isReplyFromAgent = true;
-      io.sockets.emit('wa_message', newWhatsAppMessage);
-      console.log (msg);
-      // TODO Need insert reply into database
-    }).catch(error => {
-      console.log(error);
-    });
-
-    // Twiml response
-    // TODO Commented this code, need investigate later.
-    // const response  = twilio.twiml.MessagingResponse('hello world');
-  });
+  handleSocket(socket);
 });
 /** Socket io connection */
